Rename formatNodes to formatCSS
