+++ src/lib/formatCSS.js
